requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
sqlite3-binary==0.0.1
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==2.9.2
